pyproject.toml
[project]
name = "insta-images"
version = "0.1.0"
description = ""
authors = [
    {name = "lokesh"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3 (>=1.40.43,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "jsonschema (>=4.25.1,<5.0.0)"
]

[tool.poetry]
packages = [{include = "insta_images", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "moto (>=5.1.13,<6.0.0)"
]
